FLOW-482 Now only deletes the output file if it exists (re: code review).

+++ uploadFlow.sh
@@ -36,4 +36,4 @@ else
   cat uploadFlowResponse.txt
 fi
 
-rm uploadFlowResponse.txt
+[ -e uploadFlowResponse.txt ] && rm uploadFlowResponse.txt
